IA financeiro/requirements.txt
streamlit==1.32.2
google-generativeai==0.4.1
python-dotenv==1.0.1
pandas==2.2.1
plotly==5.20.0
